pyproject.toml
[tool.poetry]
name = "useful-tools"
version = "0.1.0"
description = ""
authors = ["2lu3 <hi2lu3@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
google-cloud-texttospeech = "^2.16.3"
loguru = "^0.7.2"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
